# backend/authentication-service/src/routes/swagger_auth_service.yaml
swagger: "2.0"
info:
  title: Google & Email/Password Authentication Service API
  version: 1.2.0
  description: >
    This service handles Google OAuth2 authentication flow and simple email/password registration & login.
host: localhost:4001
basePath: /
schemes:
  - http

servers:
  - url: http://localhost:4001
    description: Local development server
    
paths:
  /auth/login:
    get:
      summary: Generate Google OAuth2 Authorization URL
      description: Returns the Google OAuth2 consent screen URL where the user will log in.
      tags:
        - Google OAuth
      responses:
        "200":
          description: Authorization URL generated
          schema:
            type: object
            properties:
              auth_url:
                type: string

  /login:
    post:
      summary: Login with email and password
      tags:
        - Email Auth
      consumes:
        - application/json
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
            required:
              - email
              - password
      responses:
        "200":
          description: Login successful (sets httpOnly access & refresh token cookies)
          schema:
            $ref: '#/definitions/EmailLoginSuccess'
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/ErrorResponse'

  /register:
    post:
      summary: Register a new user with email, password, and username
      tags:
        - Email Auth
      consumes:
        - application/json
      parameters:
        - in: body
          name: user
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              username:
                type: string
            required:
              - email
              - password
              - username
      responses:
        "201":
          description: User registered successfully. If auto-login succeeds returns tokens; otherwise returns a message.
          schema:
            $ref: '#/definitions/RegisterSuccess'
        "400":
          description: Registration failed
          schema:
            $ref: '#/definitions/ErrorResponse'

  /oauth2callback:
    get:
      summary: Handle Google OAuth2 Callback
      description: Exchanges authorization code for tokens, verifies them, sets httpOnly cookies, and redirects to frontend.
      tags:
        - Google OAuth
      parameters:
        - name: code
          in: query
          required: true
          type: string
          description: Authorization code returned by Google
      responses:
        "302":
          description: "Redirect to frontend dashboard on successful authentication (cookies set: access_token, refresh_token)"
        "400":
          description: Invalid credentials or token
          schema:
            $ref: '#/definitions/ErrorResponse'

  /auth/logout:
    post:
      summary: Log out the authenticated user
      description: Removes auth cookies (access_token, refresh_token) and returns a message.
      tags:
        - Utility
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/MessageResponse'

  /delete-user:
    delete:
      summary: Delete a user by ID
      description: Permanently removes a user from the database.
      tags:
        - Email Auth
      consumes:
        - application/json
      parameters:
        - in: body
          name: user
          required: true
          schema:
            type: object
            properties:
              user_id:
                type: integer
                example: 3
            required:
              - user_id
      responses:
        "200":
          description: User deleted
          schema:
            $ref: '#/definitions/MessageResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponse'

  /:
    get:
      summary: Health Check
      description: Check if the Authentication Service is running.
      tags:
        - Utility
      responses:
        "200":
          description: Service running successfully
          schema:
            $ref: '#/definitions/MessageResponse'

  /users:
    get:
      summary: List all users
      description: Returns an array of all registered users.
      tags:
        - Utility
      responses:
        "200":
          description: Users fetched successfully
          schema:
            type: object
            properties:
              users:
                type: array
                items:
                  $ref: '#/definitions/User'

  /auth/refresh:
    post:
      summary: Refresh access token
      description: Accepts a refresh token from cookie or request body and returns new access & refresh tokens.
      tags:
        - Utility
      consumes:
        - application/json
      parameters:
        - in: body
          name: refresh
          required: false
          schema:
            type: object
            properties:
              refresh_token:
                type: string
      responses:
        "200":
          description: Tokens refreshed successfully (sets new access_token cookie)
          schema:
            $ref: '#/definitions/TokensResponse'
        "400":
          description: Missing refresh token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/ErrorResponse'

definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        example: 1
      email:
        type: string
        example: user@example.com
      username:
        type: string
        example: johndoe
      google_id:
        type: string
        example: 108976234234234234234
      auth_provider:
        type: string
        example: google
      created_at:
        type: string
        example: 2025-11-03T10:15:00Z
  TokensResponse:
    type: object
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
        example: bearer
      user:
        $ref: '#/definitions/User'
  EmailLoginSuccess:
    type: object
    properties:
      tokens:
        $ref: '#/definitions/TokensResponse'
      email:
        type: string
        example: user@example.com
  RegisterSuccess:
    type: object
    description: Response after registration. If auto-login succeeds, tokens are present; otherwise only a message.
    properties:
      tokens:
        $ref: '#/definitions/TokensResponse'
      email:
        type: string
        example: user@example.com
      message:
        type: string
        example: Registered successfully, but login failed
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: Operation successful
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Invalid credentials